requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
faiss-cpu
pypdf
python-dotenv
sentence-transformers
